[operator] rename operator to control-operator
