+++ control-operator/PROJECT
@@ -9,7 +9,7 @@ plugins:
   manifests.sdk.operatorframework.io/v2: {}
   scorecard.sdk.operatorframework.io/v2: {}
 projectName: operator
-repo: github.com/AliceO2Group/Control/operator
+repo: github.com/AliceO2Group/Control/control-operator
 resources:
 - api:
     crdVersion: v1
@@ -18,7 +18,7 @@ resources:
   domain: alice.cern
   group: aliecs
   kind: Task
-  path: github.com/AliceO2Group/Control/operator/api/v1alpha1
+  path: github.com/AliceO2Group/Control/control-operator/api/v1alpha1
   version: v1alpha1
 - api:
     crdVersion: v1
@@ -27,7 +27,7 @@ resources:
   domain: alice.cern
   group: aliecs
   kind: Environment
-  path: github.com/AliceO2Group/Control/operator/api/v1alpha1
+  path: github.com/AliceO2Group/Control/control-operator/api/v1alpha1
   version: v1alpha1
 - api:
     crdVersion: v1
@@ -35,6 +35,6 @@ resources:
   domain: alice.cern
   group: aliecs
   kind: TaskTemplate
-  path: github.com/AliceO2Group/Control/operator/api/v1alpha1
+  path: github.com/AliceO2Group/Control/control-operator/api/v1alpha1
   version: v1alpha1
 version: "3"
