+++ control-operator/go.mod
@@ -1,4 +1,4 @@
-module github.com/AliceO2Group/Control/operator
+module github.com/AliceO2Group/Control/control-operator
 
 go 1.25.0
 
@@ -82,4 +82,4 @@ require (
 	sigs.k8s.io/yaml v1.6.0 // indirect
 )
 
-replace github.com/AliceO2Group/Control/operator/internal/controller/protos/generated => ./internal/controller/protos/generated
+replace github.com/AliceO2Group/Control/control-operator/internal/controller/protos/generated => ./internal/controller/protos/generated
